+++ mobile/app.config.js
@@ -1,4 +1,3 @@
-import "expo-share-intent/build/useSharingIntent";
 
 module.exports = ({ config }) => ({
   ...config,